+++ .vscode/settings.json
@@ -42,8 +42,10 @@
     ".firebase": true,
     ".idea": true,
     ".next": true,
+    ".sentryclirc": true,
     ".vitepress/cache": true,
     ".yarn": true,
+    "build-storybook.log": true,
     "dist": true,
     "graphql/__generated__": true,
     "next-env.d.ts": true,
