+++ app/[lang]/generation/_components/editor-task-view-view/generation-editor-task-view.tsx
@@ -3,7 +3,6 @@
 import { GenerationEditorTaskListArea } from "@/app/[lang]/generation/_components/editor-task-view-view/generation-editor-task-list-area"
 import { GenerationTaskListActions } from "@/app/[lang]/generation/_components/editor-task-view-view/generation-task-list-actions"
 import { GenerationEditorCard } from "@/app/[lang]/generation/_components/generation-editor-card"
-import { Separator } from "@/components/ui/separator"
 import { useState } from "react"
 
 /**
