+++ app/[lang]/generation/_components/generation-editor-layout.stories.tsx
@@ -17,6 +17,6 @@ export const LITE: Story = {
     negativePromptEditor: (
       <div className="bg-green-500 h-full">{"negativePromptEditor"}</div>
     ),
-    history: <div className="bg-yellow-500 h-full">{"history"}</div>,
+    taskList: <div className="bg-yellow-500 h-full">{"history"}</div>,
   },
 }
