+++ components/app/app-fixed-content.tsx
@@ -1,3 +1,5 @@
+import { cn } from "@/lib/utils"
+
 type Props = {
   position: "top" | "bottom" | "left" | "right"
   children: React.ReactNode
@@ -9,26 +11,17 @@ type Props = {
  * @returns
  */
 export const AppFixedContent = (props: Props) => {
-  /**
-   * 位置に関するクラス名を取得する
-   * @returns クラス名
-   */
-  const positionClassName = () => {
-    if (props.position === "top") {
-      return "top-0 left-0"
-    }
-    if (props.position === "bottom") {
-      return "bottom-0 left-0"
-    }
-    if (props.position === "left") {
-      return "top-0 left-0"
-    }
-    return "top-0 right-0"
-  }
-
   return (
     <div
-      className={`fixed p-2 w-full h-auto bg-card shadow-md z-50 ${positionClassName()}`}
+      className={cn(
+        "fixed container py-4 w-full h-auto bg-card shadow-md z-50",
+        "md:static md:p-0 md:bg-none md:shadow-none",
+        {
+          "top-0 left-0": props.position === "top" || props.position === "left",
+          "bottom-0 left-0": props.position === "bottom",
+          "top-0 right-0": props.position === "right",
+        },
+      )}
     >
       {props.children}
     </div>
